Move ContractClass to state in starknet_api
